Features/hmr4

+++ JavaEE_Rest_Service/TaskManager/src/main/java/com/aston/service/api/UserServiceApi.java
@@ -8,10 +8,10 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface UserServiceApi {
-    int createUser(UserDto userSave) throws SQLException, UserInvalidParameterException;
-    UserDto getUserById(int userId) throws SQLException, UserNotFoundException;
+    Long createUser(UserDto userSave) throws SQLException, UserInvalidParameterException;
+    UserDto getUserById(Long userId) throws SQLException, UserNotFoundException;
     UserDto getUserByUsername(String username) throws SQLException, UserNotFoundException;
     List<UserDto> getAllUsers() throws  SQLException;
-    int updateUser(UserDto userUpdate) throws SQLException, UserInvalidParameterException;
-    int deleteUser(int userId) throws SQLException, UserNotFoundException;
+    Long updateUser(UserDto userUpdate) throws SQLException, UserInvalidParameterException;
+    Long deleteUser(Long userId) throws SQLException, UserNotFoundException;
 }
